requirements.txt
accelerate==0.29.3
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
bitsandbytes==0.41.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.3.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==2.16.1
debugpy==1.8.17
decorator==5.2.1
decord==0.6.0
deepspeed==0.14.4
dill==0.3.7
docstring_parser==0.17.0
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.119.0
filelock==3.19.1
flash-attn==2.6.3
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2023.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
gradio_client==0.2.9
h11==0.14.0
hf-xet==1.1.10
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.35.3
idna==3.4
ipykernel==7.0.1
ipython==8.37.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
latex2mathml==3.78.1
# Editable install with no version control (llava==1.7.0.dev0)
-e /root/userfolder/MIL/VL-MIL
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
nltk==3.9.2
numpy==1.26.1
nvidia-ml-py==13.580.82
open_clip_torch==3.2.0
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.4.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==1.10.8
Pygments==2.19.2
pyparsing==3.2.5
pypinyin==0.55.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rouge==1.0.1
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.1.99
sentry-sdk==2.41.0
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.15.2
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@1c39974a4c4036fd641bc1191cc32799f85715a4
triton==2.1.0
typeguard==4.4.4
typing_extensions==4.15.0
tyro==0.9.35
tzdata==2025.2
urllib3==1.26.13
uvicorn==0.37.0
wandb==0.22.2
wavedrom==2.0.3.post3
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
